+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 + 支持路径规划：
   + AMap.Driving 驾车路线规划服务：AmapDrivingService
   + AMap.Transfer 公交换乘路线规划服务：AmapTransferService
+  + AMap.Walking 步行路线规划服务：AmapWalkingService
 
 ----